modularize onResize
