+++ src/defaultSettings.js
@@ -2,6 +2,8 @@ export default {
     //Addition settings for this template
     id_col: 'USUBJID',
     time_col: 'VISITN',
+    visit_col: 'VISIT',
+    visit_order_col: 'VISITNUM',
     measure_col: 'TEST',
     value_col: 'STRESN',
     start_value: null,
