+++ src/onDraw.js
@@ -1,10 +1,10 @@
 import updateParticipantCount from './onDraw/updateParticipantCount';
-import resetListing from './onDraw/resetListing';
+import reset from './onDraw/reset';
 
 export default function onDraw() {
     //Annotate selected and total number of participants.
     updateParticipantCount(this, '.annote');
 
-    //Reset listing.
-    resetListing.call(this);
+    //Reset things.
+    reset.call(this);
 }
