--- src/onDraw/resetListing.js
@@ -1,4 +1,0 @@
-export default function resetListing() {
-    this.listing.draw([]);
-    this.listing.wrap.style('display', 'none');
-}
